Migrates swerve module
