+++ src/main/java/frc/robot/Robot.java
@@ -7,14 +7,15 @@ package frc.robot;
 import java.util.Timer;
 import java.util.TimerTask;
 
+import com.chaos131.logging.LogManager;
+import com.chaos131.util.DashboardNumber;
+
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.RobotController;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.Constants.DebugConstants;
-import frc.robot.logging.LogManager;
-import frc.robot.util.DashboardNumber;
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
@@ -24,7 +25,7 @@ import frc.robot.util.DashboardNumber;
  */
 public class Robot extends TimedRobot {
   private Command m_autonomousCommand;
-  public static final LogManager logManager = new LogManager();
+  public static final LogManager logManager = LogManager.getInstance();
 
   private RobotContainer m_robotContainer;
