+++ src/main/java/frc/robot/commands/SwerveXMode.java
@@ -23,7 +23,7 @@ public class SwerveXMode extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_swerveDrive.swerveXMode();
+    m_swerveDrive.setXMode();
 
   }
